requirements.txt
fastapi
pillow
python-magic